Refactor: Formula metadata parsing
